table_extraction/requirements.txt
PyPDF2==3.0.1
pypdf
langchain==0.3.0
langchain_community==0.3.0
langchain_huggingface==0.1.0
langchain_core==0.3.5
langchain_postgres==0.0.12
langchain-experimental==0.3.0
torch==2.4.0
cryptography==3.1
#psycopg2==2.9.9
psycopg[binary,pool]==3.2.1
python-dotenv
datetime==5.5
psycopg-binary
langchain-openai==0.2.0
openai==1.47.0
unstructured[all-docs]==0.15.12
#flash_attn==2.6.3
pytesseract==0.3.13
